// src/proto/tyto/runtime/v1/guest.ts
export interface ExecRequest {
  start?: ExecStart | undefined;
  stdin?: StdinData | undefined;
  cancel?: ExecCancel | undefined;
  resize?: ExecResize | undefined;
}

export interface ExecResponse {
  stdout?: StdoutData | undefined;
  stderr?: StderrData | undefined;
  exit?: ExecExit | undefined;
}

export interface ExecStart {
  command: string[];
  env: { [key: string]: string };
  workingDir: string;
  tty: boolean;
  cols: number;
  rows: number;
}
// ...
export const ExecRequest: MessageFns<ExecRequest> = {
  encode(message: ExecRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.start !== undefined) {
      ExecStart.encode(message.start, writer.uint32(10).fork()).join();
    }
    if (message.stdin !== undefined) {
      StdinData.encode(message.stdin, writer.uint32(18).fork()).join();
    }
    if (message.cancel !== undefined) {
      ExecCancel.encode(message.cancel, writer.uint32(26).fork()).join();
    }
    if (message.resize !== undefined) {
      ExecResize.encode(message.resize, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ExecRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseExecRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.start = ExecStart.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.stdin = StdinData.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.cancel = ExecCancel.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.resize = ExecResize.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ExecRequest {
    return {
      start: isSet(object.start) ? ExecStart.fromJSON(object.start) : undefined,
      stdin: isSet(object.stdin) ? StdinData.fromJSON(object.stdin) : undefined,
      cancel: isSet(object.cancel) ? ExecCancel.fromJSON(object.cancel) : undefined,
      resize: isSet(object.resize) ? ExecResize.fromJSON(object.resize) : undefined,
    };
  },

  toJSON(message: ExecRequest): unknown {
    const obj: any = {};
    if (message.start !== undefined) {
      obj.start = ExecStart.toJSON(message.start);
    }
    if (message.stdin !== undefined) {
      obj.stdin = StdinData.toJSON(message.stdin);
    }
    if (message.cancel !== undefined) {
      obj.cancel = ExecCancel.toJSON(message.cancel);
    }
    if (message.resize !== undefined) {
      obj.resize = ExecResize.toJSON(message.resize);
    }
    return obj;
  },

  create(base?: DeepPartial<ExecRequest>): ExecRequest {
    return ExecRequest.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<ExecRequest>): ExecRequest {
    const message = createBaseExecRequest();
    message.start = (object.start !== undefined && object.start !== null)
      ? ExecStart.fromPartial(object.start)
      : undefined;
    message.stdin = (object.stdin !== undefined && object.stdin !== null)
      ? StdinData.fromPartial(object.stdin)
      : undefined;
    message.cancel = (object.cancel !== undefined && object.cancel !== null)
      ? ExecCancel.fromPartial(object.cancel)
      : undefined;
    message.resize = (object.resize !== undefined && object.resize !== null)
      ? ExecResize.fromPartial(object.resize)
      : undefined;
    return message;
  },
};

function createBaseExecResponse(): ExecResponse {
  return { stdout: undefined, stderr: undefined, exit: undefined };
}

export const ExecResponse: MessageFns<ExecResponse> = {
  encode(message: ExecResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.stdout !== undefined) {
      StdoutData.encode(message.stdout, writer.uint32(10).fork()).join();
    }
    if (message.stderr !== undefined) {
      StderrData.encode(message.stderr, writer.uint32(18).fork()).join();
    }
    if (message.exit !== undefined) {
      ExecExit.encode(message.exit, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ExecResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseExecResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.stdout = StdoutData.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.stderr = StderrData.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.exit = ExecExit.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ExecResponse {
    return {
      stdout: isSet(object.stdout) ? StdoutData.fromJSON(object.stdout) : undefined,
      stderr: isSet(object.stderr) ? StderrData.fromJSON(object.stderr) : undefined,
      exit: isSet(object.exit) ? ExecExit.fromJSON(object.exit) : undefined,
    };
  },

  toJSON(message: ExecResponse): unknown {
    const obj: any = {};
    if (message.stdout !== undefined) {
      obj.stdout = StdoutData.toJSON(message.stdout);
    }
    if (message.stderr !== undefined) {
      obj.stderr = StderrData.toJSON(message.stderr);
    }
    if (message.exit !== undefined) {
      obj.exit = ExecExit.toJSON(message.exit);
    }
    return obj;
  },

  create(base?: DeepPartial<ExecResponse>): ExecResponse {
    return ExecResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<ExecResponse>): ExecResponse {
    const message = createBaseExecResponse();
    message.stdout = (object.stdout !== undefined && object.stdout !== null)
      ? StdoutData.fromPartial(object.stdout)
      : undefined;
    message.stderr = (object.stderr !== undefined && object.stderr !== null)
      ? StderrData.fromPartial(object.stderr)
      : undefined;
    message.exit = (object.exit !== undefined && object.exit !== null) ? ExecExit.fromPartial(object.exit) : undefined;
    return message;
  },
};

function createBaseExecStart(): ExecStart {
  return { command: [], env: {}, workingDir: "", tty: false, cols: 0, rows: 0 };
}

export const ExecStart: MessageFns<ExecStart> = {
  encode(message: ExecStart, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.command) {
      writer.uint32(10).string(v!);
    }
    globalThis.Object.entries(message.env).forEach(([key, value]: [string, string]) => {
      ExecStart_EnvEntry.encode({ key: key as any, value }, writer.uint32(18).fork()).join();
    });
    if (message.workingDir !== "") {
      writer.uint32(26).string(message.workingDir);
    }
    if (message.tty !== false) {
      writer.uint32(32).bool(message.tty);
    }
    if (message.cols !== 0) {
      writer.uint32(40).uint32(message.cols);
    }
    if (message.rows !== 0) {
      writer.uint32(48).uint32(message.rows);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ExecStart {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseExecStart();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.command.push(reader.string());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          const entry2 = ExecStart_EnvEntry.decode(reader, reader.uint32());
          if (entry2.value !== undefined) {
            message.env[entry2.key] = entry2.value;
          }
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.workingDir = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.tty = reader.bool();
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.cols = reader.uint32();
          continue;
        }
        case 6: {
          if (tag !== 48) {
            break;
          }

          message.rows = reader.uint32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ExecStart {
    return {
      command: globalThis.Array.isArray(object?.command) ? object.command.map((e: any) => globalThis.String(e)) : [],
      env: isObject(object.env)
        ? (globalThis.Object.entries(object.env) as [string, any][]).reduce(
          (acc: { [key: string]: string }, [key, value]: [string, any]) => {
            acc[key] = globalThis.String(value);
            return acc;
          },
          {},
        )
        : {},
      workingDir: isSet(object.workingDir)
        ? globalThis.String(object.workingDir)
        : isSet(object.working_dir)
        ? globalThis.String(object.working_dir)
        : "",
      tty: isSet(object.tty) ? globalThis.Boolean(object.tty) : false,
      cols: isSet(object.cols) ? globalThis.Number(object.cols) : 0,
      rows: isSet(object.rows) ? globalThis.Number(object.rows) : 0,
    };
  },

  toJSON(message: ExecStart): unknown {
    const obj: any = {};
    if (message.command?.length) {
      obj.command = message.command;
    }
    if (message.env) {
      const entries = globalThis.Object.entries(message.env) as [string, string][];
      if (entries.length > 0) {
        obj.env = {};
        entries.forEach(([k, v]) => {
          obj.env[k] = v;
        });
      }
    }
    if (message.workingDir !== "") {
      obj.workingDir = message.workingDir;
    }
    if (message.tty !== false) {
      obj.tty = message.tty;
    }
    if (message.cols !== 0) {
      obj.cols = Math.round(message.cols);
    }
    if (message.rows !== 0) {
      obj.rows = Math.round(message.rows);
    }
    return obj;
  },

  create(base?: DeepPartial<ExecStart>): ExecStart {
    return ExecStart.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<ExecStart>): ExecStart {
    const message = createBaseExecStart();
    message.command = object.command?.map((e) => e) || [];
    message.env = (globalThis.Object.entries(object.env ?? {}) as [string, string][]).reduce(
      (acc: { [key: string]: string }, [key, value]: [string, string]) => {
        if (value !== undefined) {
          acc[key] = globalThis.String(value);
        }
        return acc;
      },
      {},
    );
    message.workingDir = object.workingDir ?? "";
    message.tty = object.tty ?? false;
    message.cols = object.cols ?? 0;
    message.rows = object.rows ?? 0;
    return message;
  },
};
// ...
export type GuestServiceService = typeof GuestServiceService;
// ...
export const GuestServiceClient = makeGenericClientConstructor(
  GuestServiceService,
  "tyto.runtime.v1.GuestService",
) as unknown as {
  new (address: string, credentials: ChannelCredentials, options?: Partial<ClientOptions>): GuestServiceClient;
  service: typeof GuestServiceService;
  serviceName: string;
};